// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Programacion/ParametrosProg.ts
export interface ParametrosProg {
  id: number;
  resistenciasAfines: boolean;
  refileMaximo: number;
  refileMinimo: number;
  diasAdelanto: number;
  todosAnchos: boolean;
  anchoCalculo: number;
  largoMinimo: number;
  excedente: number;
  scores: number;
}
export class ParametrosProg {
  id: number;
  resistenciasAfines: boolean;
  refileMaximo: number;
  refileMinimo: number;
  diasAdelanto: number;
  todosAnchos: boolean;
  anchoCalculo: number;
  largoMinimo: number;
  excedente: number;
  scores: number;
}